main-ms-interface/src/main/java/fr/shoppo/mainmsinterface/domain/bo/Adresse.java
package fr.shoppo.mainmsinterface.domain.bo;

public class Adresse {
    String codePostal;
    String numeroRue;
    String libelleRue;
    String ville;

    public Adresse() {
        /*binding*/
    }

    public String getCodePostal() {
        return codePostal;
    }

    public void setCodePostal(String codePostal) {
        this.codePostal = codePostal;
    }

    public String getNumeroRue() {
        return numeroRue;
    }

    public void setNumeroRue(String numeroRue) {
        this.numeroRue = numeroRue;
    }

    public String getLibelleRue() {
        return libelleRue;
    }

    public void setLibelleRue(String libelleRue) {
        this.libelleRue = libelleRue;
    }

    public String getVille() {
        return ville;
    }

    public void setVille(String ville) {
        this.ville = ville;
    }
}
